Flytt unsplash-relatert sqip-kode til egen fil
