+++ app/routes/unsplash/$photoId.tsx
@@ -1,10 +1,10 @@
-import type { Konvertering } from "@prisma/client";
 import { Link, useLoaderData } from "@remix-run/react";
-import type { LoaderFunction } from "@remix-run/server-runtime";
-import type { Metadata } from "../../services/sqip/sqip";
 import { json } from "@remix-run/server-runtime";
+import { fetchFromUnsplashAndRunThroughSqip } from "~/services/sqip/fraUnsplash";
 import { PhotoAttribution } from "../../components/PhotoAttribution";
-import { fetchFromUnsplashAndRunThroughSqip } from "../../services/sqip/sqip";
+import type { Konvertering } from "@prisma/client";
+import type { LoaderFunction } from "@remix-run/server-runtime";
+import type { Metadata } from "~/services/sqip/fraUnsplash";
 
 export const loader: LoaderFunction = async ({ params }) => {
     const photoId = params.photoId;
@@ -16,6 +16,7 @@ export const loader: LoaderFunction = async ({ params }) => {
     const resultatFraKonvertering = await fetchFromUnsplashAndRunThroughSqip(
         photoId
     );
+
     return json({ result: resultatFraKonvertering }, 200);
 };
